+++ demo.py
@@ -25,7 +25,7 @@ try:
     # fully empty the database first
     sampledb.utils.empty_database(sqlalchemy.create_engine(sampledb.config.SQLALCHEMY_DATABASE_URI))
     app = create_app()
-
+    app.config['SERVER_NAME'] = 'localhost:5000'
     with app.app_context():
         setup_data(app=app)