+++ sampledb/frontend/projects_forms.py
@@ -24,6 +24,18 @@ class LeaveProjectForm(FlaskForm):
     pass
 
 
+class DeleteProjectForm(FlaskForm):
+    pass
+
+
+class RemoveProjectMemberForm(FlaskForm):
+    pass
+
+
+class RemoveProjectGroupForm(FlaskForm):
+    pass
+
+
 class OtherProjectIdForm(FlaskForm):
     project_id = IntegerField(
         validators=[InputRequired()]