+++ sampledb/frontend/users/forms.py
@@ -23,6 +23,14 @@ class LeaveGroupForm(FlaskForm):
     pass
 
 
+class DeleteGroupForm(FlaskForm):
+    pass
+
+
+class RemoveGroupMemberForm(FlaskForm):
+    pass
+
+
 class InviteUserForm(FlaskForm):
     user_id = IntegerField(validators=[InputRequired()])